Small Internet-connection parser for data from VK page using wininet.h
There is a prog, that allows you to get an information about personal VK
page, by typing it's id to file "id.txt" (that file must be required).
Also Internet connection required. Represent result as an array of
structures, so it's possible to set the id list in file (example
included).